Transfer engine to gc and fixes
