+++ boa_engine/Cargo.toml
@@ -37,7 +37,6 @@ boa_profiler.workspace = true
 boa_macros.workspace = true
 boa_ast.workspace = true
 boa_parser.workspace = true
-gc = "0.4.1"
 serde = { version = "1.0.147", features = ["derive", "rc"] }
 serde_json = "1.0.87"
 rand = "0.8.5"
