+++ boa_engine/src/builtins/promise/mod.rs
@@ -21,7 +21,7 @@ use crate::{
     value::JsValue,
     Context, JsError, JsResult,
 };
-use boa_gc::{Cell as GcCell, Finalize, Gc, Trace};
+use boa_gc::{BoaAlloc, Cell as GcCell, Finalize, Gc, Trace};
 use boa_profiler::Profiler;
 use std::{cell::Cell, rc::Rc};
 use tap::{Conv, Pipe};
@@ -118,10 +118,10 @@ impl PromiseCapability {
 
                 // 2. NOTE: C is assumed to be a constructor function that supports the parameter conventions of the Promise constructor (see 27.2.3.1).
                 // 3. Let promiseCapability be the PromiseCapability Record { [[Promise]]: undefined, [[Resolve]]: undefined, [[Reject]]: undefined }.
-                let promise_capability = Gc::new(boa_gc::Cell::new(RejectResolve {
+                let promise_capability = BoaAlloc::new_cell(RejectResolve {
                     reject: JsValue::undefined(),
                     resolve: JsValue::undefined(),
-                }));
+                });
 
                 // 4. Let executorClosure be a new Abstract Closure with parameters (resolve, reject) that captures promiseCapability and performs the following steps when called:
                 // 5. Let executor be CreateBuiltinFunction(executorClosure, 2, "", « »).
@@ -436,7 +436,7 @@ impl Promise {
         }
 
         // 1. Let values be a new empty List.
-        let values = Gc::new(GcCell::new(Vec::new()));
+        let values = BoaAlloc::new_cell(Vec::new());
 
         // 2. Let remainingElementsCount be the Record { [[Value]]: 1 }.
         let remaining_elements_count = Rc::new(Cell::new(1));
@@ -677,7 +677,7 @@ impl Promise {
         }
 
         // 1. Let values be a new empty List.
-        let values = Gc::new(GcCell::new(Vec::new()));
+        let values = BoaAlloc::new_cell(Vec::new());
 
         // 2. Let remainingElementsCount be the Record { [[Value]]: 1 }.
         let remaining_elements_count = Rc::new(Cell::new(1));
@@ -1008,7 +1008,7 @@ impl Promise {
         }
 
         // 1. Let errors be a new empty List.
-        let errors = Gc::new(GcCell::new(Vec::new()));
+        let errors = BoaAlloc::new_cell(Vec::new());
 
         // 2. Let remainingElementsCount be the Record { [[Value]]: 1 }.
         let remaining_elements_count = Rc::new(Cell::new(1));
