+++ boa_engine/src/bytecompiler/function.rs
@@ -7,7 +7,7 @@ use crate::{
 use boa_ast::{
     declaration::Binding, function::FormalParameterList, operations::bound_names, StatementList,
 };
-use boa_gc::Gc;
+use boa_gc::{Gc, BoaAlloc};
 use boa_interner::Sym;
 use rustc_hash::FxHashMap;
 
@@ -219,6 +219,6 @@ impl FunctionCompiler {
         compiler.emit(Opcode::PushUndefined, &[]);
         compiler.emit(Opcode::Return, &[]);
 
-        Ok(Gc::new(compiler.finish()))
+        Ok(BoaAlloc::new(compiler.finish()))
     }
 }
