+++ boa_gc/src/internals/mod.rs
@@ -1,8 +1,6 @@
 pub(crate) mod borrow_flag;
 pub mod cell;
-pub mod cell_ref;
 pub mod ephemeron;
 
-pub use cell::GcCell;
-pub use cell_ref::{GcCellRef, GcCellRefMut};
+pub use cell::*;
 pub use ephemeron::Ephemeron;
